PACT-549 | EPR | Creating an sbt command for pre-push hook.
A new SBT command has been created, after adding sbt-git as plugin, that will take care of formating code and then adding and commiting changes into git. The purpose of this command is to be used as part of a pre-push git hook in order to make sure formatting is applied to all files before checking in code to remote.